==> globalprep/prs_land-based_nutrient/v2023/STEP1_crop/step3_fao_production_wrangling.Rmd <==
---
title: "OHI 2022: Extract and wrangle FAO stat nutrient data"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 1
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '../../../workflow/templates/ohi_hdr.html'
pdf_document:
  toc: true
editor_options: 
  chunk_output_type: console
---

# Summary
Wrangle FAOSTAT data; allocate data nationally based on crop production at each location.
https://github.com/OHI-Science/food_systems/blob/master/STEP1_crop/nutrient/step3_fao_production_wrangling.Rmd

2/3 of fertilizer use allocated to irrigated systems and 1/3 to high input systems.

***

# Data Source

**Reference**: FAOSTAT; http://www.fao.org/faostat/en/#data/RFN 

**Downloaded**: 08/26/2022

**Description**: 
The Fertilizers by Nutrient dataset contains information on the totals in nutrients for Production, Trade and Agriculture Use of inorganic (chemical or mineral) fertilizers, over the time series 1961-present. The data are provided for the three primary plant nutrients: nitrogen (N), phosphorus (expressed as P2O5) and potassium (expressed as K2O). Both straight and compound fertilizers are included.

**Native data resolution**: Country 

**Time range**: 1961-2020

**Format**: CSV

**Notes**: Go here: http://www.fao.org/faostat/en/#data/RFN  and click "Bulk Downloads" ---> "All Data", and save on Mazu.


***

# Setup
```{r setup, include = FALSE}

# getting packages we want
library(here)
library(janitor)
library(vroom)
library(cowplot)
library(tidyverse)
library(countrycode)

# Insall OHI core
#devtools::install_github('ohi-science/ohicore@dev')

# Source functions
source(here('workflow/R/common.R'))

# Update scenario year, set up programmatic scenario year updating
scen_year_number <- 2023
scen_year <- as.character(scen_year_number)
prev_scen_year <- as.character(scen_year_number - 1)

# Set up file path
crop_nutrient <- file.path(here(), paste0("globalprep/prs_land-based_nutrient/v", scen_year, "/STEP1_crop"))

# Load OHI data
region_data()

# Load food regions, this does not update yearly
food_rgns <- read_csv(here("globalprep/spatial/v2021/food_rgns.csv"), col_types = "cdc")

# Load in FAO data
nutrient_anx <- file.path(paste0("/home/shares/ohi/git-annex/globalprep/_raw_data/FAOSTAT/crop_nutrient/d", scen_year))

``` 

# Methods 

## Integrate iso3c codes
```{r}

# Import country-level fertilizer data
fao_nutrient <-   vroom::vroom(file.path(nutrient_anx, "Inputs_FertilizersNutrient_E_All_Data_NOFLAG.csv"), 
               .name_repair = make_clean_names) %>%
  pivot_longer(cols = c(9:68), names_to = "year") %>%
  mutate(year = as.numeric(gsub("y", "", year))) %>%
  rename(nutrient = item, country_application = value) %>%
    mutate(nutrient = recode(nutrient,
                           "Nutrient nitrogen N (total)"     = "N",
                           "Nutrient phosphate P2O5 (total)" = "P2O5",
                           "Nutrient potash K2O (total)"     = "K2O")) %>%
    filter(element == "Agricultural Use") %>%
  filter(year >= 2005) %>%
  filter(area_code < 5000)

# Global tonnes of each fertilizer type
sum_before_wrangling <- 
  fao_nutrient %>% 
  group_by(nutrient, year) %>% 
  summarize(global_tonnes = sum(country_application, na.rm = TRUE))

```

# Match FAOSTAT regions with food_rgns list
```{r}

# Fix Sudan
fix_sudan <- fao_nutrient %>%
  filter(str_detect(area, "Sudan")) %>%
  filter(!is.na(country_application)) %>%
  mutate(area = "Sudan",
         area_code = 276) %>%
  arrange(nutrient, year)

# Remove China and re-allocate Serbia data to Kosovo and Serbia and filter out serbia and montenegro from 2005 (when they split)
fao_nutrient <- fao_nutrient %>%
  mutate(area_code = if_else(area_code %in% 272, 275, area_code)) %>%  # XKO
  bind_rows(fao_nutrient) %>%
  mutate(area_code = if_else(area_code %in% 272, 286, area_code)) %>%
  unique() %>% 
  filter(area_code != 351) %>% # Code 351 values are for all of China, including MAC, HKG, TWN.
  filter(area_code != 186) %>%
  filter(!str_detect(area, "Sudan")) %>%
  rbind(fix_sudan) %>%
  filter(!(area %in% c("USSR", "Yugoslav SFR", "Ethiopia PDR", "Czechoslovakia", "Belgium-Luxembourg")))


```
  
# Add iso3c column and split serbia data into serbia and kosovo
```{r}  
# Convert fao rgns to iso3c
fao_rgn_codes <- vroom::vroom(here("globalprep/spatial/v2021/FAO_rgn_codes.csv"))

fao_nutrient <- 
  left_join(fao_nutrient, fao_rgn_codes, by = "area_code") %>% 
  dplyr::select(nutrient, year, country_application, iso3c) %>% 
  unique()  # perfect.. now there are no NAs in iso3c.

serb_mont_xko_2005_fix <- fao_nutrient %>%
  filter(iso3c %in% c("XKO", "MNE", "SRB")) %>%
  fill(country_application, .direction = "up")

fao_nutrient <- fao_nutrient %>%
  filter(!(iso3c %in% c("XKO", "MNE", "SRB"))) %>%
  rbind(serb_mont_xko_2005_fix)

# Allocate fertilizer data to Serbia and Kosovo

## read in the crop production data and get a proportional estimate of how much production to total production they have
SRB_XKO <- 
  vroom::vroom(here(crop_nutrient, "data/prod_crop_rgns_2017.csv")) %>% 
  filter(iso3c %in% c("XKO", "SRB") & prod_system %in% "A") %>% 
  group_by(iso3c) %>% 
  summarize(production = sum(production)) %>% 
  ungroup() %>% 
  mutate(relative_proportion = production / sum(production))


## Now assign those proportional values to serbia and kosovo fertilizer application
fao_nutrient$country_application[fao_nutrient$iso3c %in% "SRB"] <- 
  fao_nutrient$country_application[fao_nutrient$iso3c %in% "SRB"] * 
  SRB_XKO$relative_proportion[SRB_XKO$iso3c %in% "SRB"]

fao_nutrient$country_application[fao_nutrient$iso3c %in% "XKO"] <- 
  fao_nutrient$country_application[fao_nutrient$iso3c %in% "XKO"] * 
  SRB_XKO$relative_proportion[SRB_XKO$iso3c %in% "XKO"]
```

## Fill data for missing_rgns
```{r}
# Extract iso3c codes not included in fao_nutrient dataset
missing_rgns <- setdiff(unique(food_rgns$iso3c), unique(fao_nutrient$iso3c))
duplicated(fao_rgn_codes$area_code)
length(missing_rgns) # 74 missing rgns

# I gave these autonomous regions the same area_codes as POR and GBR respectively.
fao_nutrient$country_application[fao_nutrient$iso3c %in% c("XMI", "GGY")] <- 0

# Calculate global tonnes of each fertilizer
sum_before_allocating <- 
  fao_nutrient %>% 
  group_by(nutrient, year) %>% 
  summarize(global_tonnes_before = sum(country_application, na.rm = TRUE))
sum_before_allocating

fao_nutrient <- 
  expand_grid(iso3c = missing_rgns,
              nutrient = c("N", "P2O5", "K2O"),
              country_application = NA,
              reported_by_fao = "No", 
              year = c(2005:2020)) %>% 
  bind_rows(fao_nutrient) %>% 
  mutate(reported_by_fao = if_else(is.na(reported_by_fao),
                                   "Yes",
                                   reported_by_fao))

```

# Gapfill

```{r}
prod_crop_rgns <-
  vroom(here(crop_nutrient, "data/prod_crop_rgns_2017.csv")) %>%
  filter(prod_system %in% c("I", "H")) %>%
  mutate(production = if_else(prod_system == "I",
                              production * 2,
                              production)) %>%
  group_by(iso3c) %>%
  summarise(production = sum(production, na.rm = TRUE)) %>%
  ungroup() %>%
  arrange(desc(production))

reported_countries <-
  fao_nutrient %>%
  dplyr::select(iso3c, reported_by_fao, year, country_application, nutrient) %>%
  unique()

split_df <-
  left_join(prod_crop_rgns, reported_countries) %>%
  group_split(nutrient, year)

gapfill_nutrient <- function(grouped_df) {
  
  #grouped_df <- split_df[1]
  
  grouped_df <- rbind.data.frame(grouped_df)

  grouped_df$nearest <- grouped_df$iso3c

  ino <- grouped_df$reported_by_fao == 'No'

  iyes <- grouped_df$reported_by_fao == 'Yes'

  grouped_df$nearest[ino] <-
    sapply(grouped_df$production[ino], function(x) {
         val <- grouped_df$production[iyes]
         grouped_df$country_application[iyes][which.min(abs(val -x))]
    
    })

  grouped_df2 <-
    grouped_df %>%
    mutate(nearest = as.double(nearest)) %>%
    mutate(country_application_check = if_else(reported_by_fao == "No",
                                              nearest,
                                              country_application))

#}
  return(grouped_df)

  }


## we'll save a gapfilled version and a non-gapfilled version
  fao_nutrient_gf <- 
    lapply(split_df, gapfill_nutrient) %>%
    bind_rows() %>%
    mutate(country_application_final = as.numeric(ifelse(is.na(country_application), nearest, country_application))) %>%
    mutate(method = ifelse(is.na(country_application) & reported_by_fao == "No" & !is.na(country_application_final), "nearest use", "none")) %>%
    mutate(gapfilled = ifelse(method == "nearest use", 1, 0)) %>%
    dplyr::select(-country_application) %>%
    rename("country_application" = "country_application_final")
  
  
  fao_nutrient_non_gf <-
    fao_nutrient

# fao_nutrient_test <-
#   left_join(test, prod_crop_rgns) %>%
#   dplyr::select(- production)
  

```

# Remove the proportion of fertilizer used for grasslands from the FAO data.

```{r}
fubc_grass <- vroom(here(crop_nutrient, "int/fubc_2017_grassland.csv"))
  
fao_nutrient_adjusted_non_gf <- 
  left_join(fao_nutrient, fubc_grass) %>% 
  mutate(country_application = country_application - (country_application * grassland)) %>% 
  dplyr::select(- grassland)

fao_nutrient_adjusted_gf <- 
  left_join(fao_nutrient_gf, fubc_grass) %>% 
  mutate(country_application = country_application - (country_application * grassland)) %>% 
  dplyr::select(- grassland)

fao_nutrient_grassland <- 
  left_join(fao_nutrient, fubc_grass) %>% 
  mutate(country_application = country_application * grassland) %>% 
  dplyr::select(- grassland)
```

# Save

```{r}
# save non-gapfilled adjusted
write_csv(fao_nutrient_adjusted_non_gf, here(crop_nutrient, "int/fao_nutrient_application_no_gf.csv"))

# save gapfilled adjusted
write_csv(fao_nutrient_adjusted_gf, here(crop_nutrient, "int/fao_nutrient_application_gf.csv"))

# save grasslands adjusted
write_csv(fao_nutrient_grassland, here(crop_nutrient, "int/fao_nutrient_grassland_1.csv"))
```
